feat(drag-and-drop): implement drag-and-drop support for label reordering and group switching
